fix: add cert walkback for toucan credits
